+++ @xen-orchestra/lite/src/pages/network/new.vue
@@ -71,8 +71,8 @@ async function createNetwork(newPayload: NewNetworkPayload) {
     }
 
     redirectAfterSuccess(promiseCreateResult.value)
-  } catch (_error) {
-    error.value = _error as Error
+  } catch (rawError) {
+    error.value = rawError as Error
   }
 }