+++ @xen-orchestra/lite/src/route-map.d.ts
@@ -120,6 +120,13 @@ declare module 'vue-router/auto-routes' {
       Record<never, never>,
       | never
     >,
+    '/network/new-internal': RouteRecordInfo<
+      '/network/new-internal',
+      '/network/new-internal',
+      Record<never, never>,
+      Record<never, never>,
+      | never
+    >,
     '/object-not-found': RouteRecordInfo<
       '/object-not-found',
       '/object-not-found',
@@ -243,6 +250,13 @@ declare module 'vue-router/auto-routes' {
     '/story/linear-chart': RouteRecordInfo<
       '/story/linear-chart',
       '/story/linear-chart',
+      Record<never, never>,
+      Record<never, never>,
+      | never
+    >,
+    '/story/overlay': RouteRecordInfo<
+      '/story/overlay',
+      '/story/overlay',
       Record<never, never>,
       Record<never, never>,
       | never
@@ -397,6 +411,27 @@ declare module 'vue-router/auto-routes' {
     '/story/web-core/object-icon/vts-object-icon': RouteRecordInfo<
       '/story/web-core/object-icon/vts-object-icon',
       '/story/web-core/object-icon/vts-object-icon',
+      Record<never, never>,
+      Record<never, never>,
+      | never
+    >,
+    '/story/web-core/overlay/vts-overlay-button': RouteRecordInfo<
+      '/story/web-core/overlay/vts-overlay-button',
+      '/story/web-core/overlay/vts-overlay-button',
+      Record<never, never>,
+      Record<never, never>,
+      | never
+    >,
+    '/story/web-core/overlay/vts-overlay-cancel-button': RouteRecordInfo<
+      '/story/web-core/overlay/vts-overlay-cancel-button',
+      '/story/web-core/overlay/vts-overlay-cancel-button',
+      Record<never, never>,
+      Record<never, never>,
+      | never
+    >,
+    '/story/web-core/overlay/vts-overlay-confirm-button': RouteRecordInfo<
+      '/story/web-core/overlay/vts-overlay-confirm-button',
+      '/story/web-core/overlay/vts-overlay-confirm-button',
       Record<never, never>,
       Record<never, never>,
       | never
@@ -1108,6 +1143,12 @@ declare module 'vue-router/auto-routes' {
       views:
         | never
     }
+    'src/pages/network/new-internal.vue': {
+      routes:
+        | '/network/new-internal'
+      views:
+        | never
+    }
     'src/pages/object-not-found.vue': {
       routes:
         | '/object-not-found'
@@ -1216,6 +1257,12 @@ declare module 'vue-router/auto-routes' {
     'src/stories/linear-chart.story.vue': {
       routes:
         | '/story/linear-chart'
+      views:
+        | never
+    }
+    'src/stories/overlay.story.vue': {
+      routes:
+        | '/story/overlay'
       views:
         | never
     }
@@ -1348,6 +1395,24 @@ declare module 'vue-router/auto-routes' {
     'src/stories/web-core/object-icon/vts-object-icon.story.vue': {
       routes:
         | '/story/web-core/object-icon/vts-object-icon'
+      views:
+        | never
+    }
+    'src/stories/web-core/overlay/vts-overlay-button.story.vue': {
+      routes:
+        | '/story/web-core/overlay/vts-overlay-button'
+      views:
+        | never
+    }
+    'src/stories/web-core/overlay/vts-overlay-cancel-button.story.vue': {
+      routes:
+        | '/story/web-core/overlay/vts-overlay-cancel-button'
+      views:
+        | never
+    }
+    'src/stories/web-core/overlay/vts-overlay-confirm-button.story.vue': {
+      routes:
+        | '/story/web-core/overlay/vts-overlay-confirm-button'
       views:
         | never
     }